Replace fixed task values with simulated diminishing-returns curves so later solvers can allocate a variable token amount per task.

+++ src/allocator/allocation.py
@@ -18,5 +18,5 @@ class Allocation:
         return sum(task.token_cost for task in self.selected)
 
     @property
-    def total_value(self) -> int:
+    def total_value(self) -> float:
         return sum(task.value for task in self.selected)
